requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers
django-getenv
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.4
injector==0.21.0
python-dotenv